Refactoring level progress display service

+++ Assets/Scripts/UI/Game/EnemiesCountView.cs
@@ -11,7 +11,7 @@ namespace UI.Game
         public void Show() => gameObject.SetActive(true);
         public void Hide() => gameObject.SetActive(false);
 
-        public void Init(float enemiesDestroyedCount, float enemiesCount)
+        public void Init(int enemiesDestroyedCount, int enemiesCount)
         {
             EnemiesDestroyedCount.Init(enemiesDestroyedCount.ToString());
             EnemiesCount.Init(enemiesCount.ToString());
